Pick up item 'units' if present
Use the units property of a carrier item to display weight-loads for the carrier, and the units property of contained items to display the items' weight.

Note: there is no conversion for the total weight-load if an item is transferred between carriers having different units. That might be a nice-to-have, but it is beyond the scope of this fix.